refactor(vfs): 将do_open函数提取到open_utils模块
将sys_open.rs中的do_open函数提取到新建的open_utils模块，并在多处调用处更新引用路径。
