+++ kernel/src/filesystem/vfs/syscall/sys_open.rs
@@ -3,12 +3,8 @@
 use system_error::SystemError;
 
 use crate::arch::syscall::nr::SYS_OPEN;
-use crate::filesystem::vfs::fcntl::AtFlags;
-use crate::filesystem::vfs::open::do_sys_open;
-use crate::filesystem::vfs::{FileMode, ModeType, MAX_PATHLEN};
 use crate::syscall::table::FormattedSyscallParam;
 use crate::syscall::table::Syscall;
-use crate::syscall::user_access::check_and_clone_cstr;
 
 use alloc::string::ToString;
 use alloc::vec::Vec;
@@ -28,7 +24,7 @@ impl Syscall for SysOpenHandle {
         let flags = Self::flags(args);
         let mode = Self::mode(args);
 
-        do_open(path, flags, mode, true)
+        super::open_utils::do_open(path, flags, mode, true)
     }
 
     /// Formats the syscall arguments for display/debugging purposes.
@@ -59,34 +55,3 @@ impl SysOpenHandle {
 }
 
 syscall_table_macros::declare_syscall!(SYS_OPEN, SysOpenHandle);
-
-/// Performs the actual file opening operation.
-///
-/// # Arguments
-/// * `path` - Pointer to the path string
-/// * `o_flags` - File opening flags
-/// * `mode` - File mode/permissions
-/// * `follow_symlink` - Whether to follow symbolic links
-///
-/// # Returns
-/// File descriptor on success, or error code on failure.
-pub(super) fn do_open(
-    path: *const u8,
-    o_flags: u32,
-    mode: u32,
-    follow_symlink: bool,
-) -> Result<usize, SystemError> {
-    let path = check_and_clone_cstr(path, Some(MAX_PATHLEN))?
-        .into_string()
-        .map_err(|_| SystemError::EINVAL)?;
-
-    let open_flags: FileMode = FileMode::from_bits(o_flags).ok_or(SystemError::EINVAL)?;
-    let mode = ModeType::from_bits(mode).ok_or(SystemError::EINVAL)?;
-    return do_sys_open(
-        AtFlags::AT_FDCWD.bits(),
-        &path,
-        open_flags,
-        mode,
-        follow_symlink,
-    );
-}
